Added Rally Script, HeroicEffectController Script, Evil Script, and Noble Script
