+++ Assets/Prefabs/HeroicCard.prefab
@@ -1158,7 +1158,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6325016447726898180}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b67b42b22a189394db4481aa539c4963, type: 3}
+  m_Script: {fileID: 11500000, guid: f457bcab006e4f648aff445f21ec1fdd, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &3374166258178011545
