+++ Assets/Scripts/Card Scripts/Heroic Card Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b4221b129fc3f54aa600dc70610fefc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
