+++ Assets/Scripts/Card Scripts/Heroic Card Effects/Evil.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b67b42b22a189394db4481aa539c4963
+guid: c41400d8bd10782458b377a4c7398e82
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
