+++ Assets/Scripts/Card Scripts/Heroic Card Effects/Noble.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c5171e1715a5834eaf99fd471cc307a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
