+++ Assets/Scripts/Card Scripts/HeroicCardController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f457bcab006e4f648aff445f21ec1fdd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
