+++ Assets/Scripts/Card Scripts/Utility Card Effect Scripts/Rally.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9e8cdb5f8b52694a8e450d48cbf1269
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
